75% fewer Symbol#to_s calls in merge_attributes
In awsm, create_current_user ends up calling Symbol#to_s 1.6 million times
per invocation. 1.4 million of those come from Cistern's merge_attributes.

This patch converts the incoming key to a string and a symbol variant, once
each, and only uses those for its work. On average each key version of the
symbol was being converted to a string again 4 times in this method.

Awsm travis runs with and without:

  with optimization: ran for ran for 14 hrs 48 min 31 sec
  https://magnum.travis-ci.com/engineyard/awsm/builds/7465897

  without optimization: ran for 17 hrs 9 min 48 sec
  https://magnum.travis-ci.com/engineyard/awsm/builds/7463787

So, ~86% of the time to run whole awsm suite with this change.