style: add rubocop
This involved a huge amount of changes and some of them were functional.
Specifically there were issues with frozen string modifications with the
formatters that I could not quite fix but there was no reason I could
see to modify the strings in place so I worked around it.

This will need some integration testing before merging in.